+++ lib/grid/hexagon_offset_grid.dart
@@ -30,7 +30,7 @@ class HexagonOffsetGrid extends StatelessWidget {
   final int columns;
   final int rows;
   final Color color;
-  final EdgeInsetsGeometry hexagonPadding;
+  final double hexagonPadding;
   final Widget Function(int col, int row) buildChild;
   final HexagonWidget Function(int col, int row) buildHexagon;
 
@@ -42,7 +42,7 @@ class HexagonOffsetGrid extends StatelessWidget {
   ///
   /// [color] - Background color of this grid.
   ///
-  /// [hexagonPadding] - Used for padding around all Hexagon tiles.
+  /// [hexagonPadding] - Used for padding around all hexagon tiles. Will be overridden by template padding.
   ///
   /// [buildHexagon] - Provide a HexagonWidget.template() to be used for given tile (col,row). Returning null value will be represented as translucent tile.
   ///
@@ -53,7 +53,7 @@ class HexagonOffsetGrid extends StatelessWidget {
     this.color,
     this.buildChild,
     this.buildHexagon,
-    this.hexagonPadding: const EdgeInsets.all(0.0),
+    this.hexagonPadding,
   })  : assert(columns > 0),
         assert(rows > 0),
         this.hexType = HexagonType.FLAT,
@@ -67,7 +67,7 @@ class HexagonOffsetGrid extends StatelessWidget {
   ///
   /// [color] - Background color of this grid.
   ///
-  /// [hexagonPadding] - Used for padding around all Hexagon tiles.
+  /// [hexagonPadding] - Used for padding around all hexagon tiles. Will be overridden by template padding.
   ///
   /// [buildHexagon] - Provide a HexagonWidget.template() to be used for given tile (col,row). Returning null value will be represented as translucent tile.
   ///
@@ -78,7 +78,7 @@ class HexagonOffsetGrid extends StatelessWidget {
     this.color,
     this.buildChild,
     this.buildHexagon,
-    this.hexagonPadding: const EdgeInsets.all(0.0),
+    this.hexagonPadding,
   })  : this.hexType = HexagonType.FLAT,
         this.gridType = GridType.EVEN;
 
@@ -90,7 +90,7 @@ class HexagonOffsetGrid extends StatelessWidget {
   ///
   /// [color] - Background color of this grid.
   ///
-  /// [hexagonPadding] - Used for padding around all Hexagon tiles.
+  /// [hexagonPadding] - Used for padding around all hexagon tiles. Will be overridden by template padding.
   ///
   /// [buildHexagon] - Provide a HexagonWidget.template() to be used for given tile (col,row). Returning null value will be represented as translucent tile.
   ///
@@ -101,7 +101,7 @@ class HexagonOffsetGrid extends StatelessWidget {
     this.color,
     this.buildChild,
     this.buildHexagon,
-    this.hexagonPadding: const EdgeInsets.all(0.0),
+    this.hexagonPadding,
   })  : this.hexType = HexagonType.POINTY,
         this.gridType = GridType.ODD;
 
@@ -113,7 +113,7 @@ class HexagonOffsetGrid extends StatelessWidget {
   ///
   /// [color] - Background color of this grid.
   ///
-  /// [hexagonPadding] - Used for padding around all Hexagon tiles.
+  /// [hexagonPadding] - Used for padding around all hexagon tiles. Will be overridden by template padding.
   ///
   /// [buildHexagon] - Provide a HexagonWidget.template() to be used for given tile (col,row). Returning null value will be represented as translucent tile.
   ///
@@ -124,7 +124,7 @@ class HexagonOffsetGrid extends StatelessWidget {
     this.color,
     this.buildChild,
     this.buildHexagon,
-    this.hexagonPadding: const EdgeInsets.all(0.0),
+    this.hexagonPadding,
   })  : this.hexType = HexagonType.POINTY,
         this.gridType = GridType.EVEN;
 
@@ -149,11 +149,13 @@ class HexagonOffsetGrid extends StatelessWidget {
                 (crossCount) => List.generate(crossCount, (crossIndex) {
                   if ((crossIndex == 0 || crossIndex >= crossCount - 1) &&
                       gridType.displace(mainIndex, crossIndex)) {
+                    //return container with half the size of the hexagon for displaced row/column
                     return Container(
                       width: hexType.isPointy ? size.width / 2 : null,
                       height: hexType.isFlat ? size.height / 2 : null,
                     );
                   }
+                  //calculate human readable column & row
                   final col = (hexType.isPointy
                       ? (crossIndex -
                           (gridType.displaceFront(mainIndex) ? 1 : 0))
@@ -166,22 +168,23 @@ class HexagonOffsetGrid extends StatelessWidget {
                   var hexagonTemplate = buildHexagon?.call(col, row);
 
                   if (buildHexagon != null && hexagonTemplate == null) {
+                    //create default template when no build function or nothing generated
                     hexagonTemplate =
                         HexagonWidget.template(color: Colors.transparent);
-                  } else if (hexagonTemplate?.isTemplate != true)
+                  } else if (hexagonTemplate?.isTemplate != true) {
+                    //if null or not a template then override with default template
                     hexagonTemplate = HexagonWidget.template();
-
-                  return Padding(
+                  }
+                  //use template values
+                  return hexagonTemplate?.copyWith(
+                    type: hexType,
+                    inBounds: false,
                     padding: hexagonPadding,
-                    child: hexagonTemplate?.copyWith(
-                      type: hexType,
-                      inBounds: false,
-                      width: hexType.isPointy ? size.width : null,
-                      height: hexType.isFlat ? size.height : null,
-                      child: buildChild != null
-                          ? buildChild.call(col, row)
-                          : hexagonTemplate.child,
-                    ),
+                    width: hexType.isPointy ? size.width : null,
+                    height: hexType.isFlat ? size.height : null,
+                    child: buildChild != null
+                        ? buildChild.call(col, row)
+                        : hexagonTemplate.child,
                   );
                 }),
               ),
@@ -213,23 +216,19 @@ class HexagonOffsetGrid extends StatelessWidget {
   Size _hexSize(double maxWidth, double maxHeight) {
     if (maxWidth.isFinite) {
       if (hexType.isFlat) {
-        var quarters = (maxWidth - columns * hexagonPadding.horizontal) /
-            (1 + (0.75 * (columns - 1)));
+        var quarters = maxWidth / (1 + (0.75 * (columns - 1)));
         var size = Size(quarters, quarters * hexType.ratio);
         return size * hexType.flatFactor(false);
       }
-      var half = (maxWidth - columns * hexagonPadding.horizontal) /
-          (columns * 2 + _displaceColumns);
+      var half = maxWidth / (columns * 2 + _displaceColumns);
       return Size(half * 2, half * 2 * hexType.ratio);
     } else if (maxHeight.isFinite) {
       if (hexType.isPointy) {
-        var quarters = (maxHeight - rows * hexagonPadding.vertical) /
-            (1 + (0.75 * (rows - 1)));
+        var quarters = maxHeight / (1 + (0.75 * (rows - 1)));
         var size = Size(quarters / hexType.ratio, quarters);
         return size * hexType.pointyFactor(false);
       }
-      var half = (maxHeight - rows * hexagonPadding.vertical) /
-          (rows * 2 + _displaceRows);
+      var half = (maxHeight - 0) / (rows * 2 + _displaceRows);
       return Size(half * 2 / hexType.ratio, half * 2);
     } else {
       throw Exception('Error: Infinite constraints in both dimensions!');